feature: рефакторинг movies
