+++ src/components/Movies/MoviesCard.jsx
@@ -13,25 +13,8 @@ function MoviesCard({
   movieId,
   image,
   thumbnail,
-  addCard,
 }) {
   let location = useLocation(); // переменная для useLocation
-
-  function handleSave() {
-    addCard({
-      country,
-      director,
-      duration,
-      year,
-      description,
-      trailerLink,
-      nameRU,
-      nameEN,
-      movieId,
-      image,
-      thumbnail,
-    });
-  }
 
   // проверка useLocation
   function isSavedLocation() {
@@ -47,7 +30,6 @@ function MoviesCard({
         </div>
         <div className="card__save">
           <button
-            onClick={handleSave}
             className={
               isSavedLocation() ? 'card__btn card__btn_inSaved' : 'card__btn'
             }
